Add access control to postHandler and deleteHandler

+++ routes/resource.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const router = express.Router();
 var mongoose = require('mongoose');
+const ACL = require('../acl/property_access');
+const AclConfig = require('../acl/acl_config');
 
 const fieldsHelper = require('../models/fields/helper');
 
@@ -23,40 +25,49 @@ const doUpdate = (model, processor, update, res) => {
 }
 
 const postHandler = (model, processor) => (req, res) => {
-  // if it submits form data with files, the result
-  // will be encoded in payload field of req.body
-  const payload = req.body.payload ? JSON.parse(req.body.payload) : req.body;
-  const validator = fieldsHelper.getValidator(model.collection.collectionName);
-  const { errors, isValid } = validator(payload);
-  if (!isValid) {
-    return res.status(400).json(errors);
-  }
-
-  const fields = fieldsHelper.getFields(model.collection.collectionName);
-  const update = {
-    ...processor.preProcess(fields, payload, req.files),
-    updated_at: new Date()
-  };
-
-  if (!payload._id) {
-    if (processor.checkExistence) {
-      model.findOne(processor.checkExistence(req)).then(record => {
-        if (record) {
-          const name = model.collection.collectionName.slice(0, -1);
-          res.status(400).json({errorReason: name + " already exists"});
-        } else {
-          doInsert(model, processor, update, res);
-        }
-      })
-    } else {
-      doInsert(model, processor, update, res);
+  const userId = req.user.id;
+  const propertyType = model.collection.collectionName;
+  const propertyId = payload._id;
+  const role = !payload._id ? AclConfig.CreateRole : "read_write";
+  ACL.checkAccess(userId, propertyId, propertyType, role, (isAccessible) => {
+    if (!isAccessible) {
+      return res.status(401).json({err: 'No permission to update.'});
     }
-  }
-  // to update
-  else {
-    update._id = payload._id;
-    doUpdate(model, processor, update, res);
-  }
+    // if it submits form data with files, the result
+    // will be encoded in payload field of req.body
+    const payload = req.body.payload ? JSON.parse(req.body.payload) : req.body;
+    const validator = fieldsHelper.getValidator(model.collection.collectionName);
+    const { errors, isValid } = validator(payload);
+    if (!isValid) {
+      return res.status(400).json(errors);
+    }
+
+    const fields = fieldsHelper.getFields(model.collection.collectionName);
+    const update = {
+      ...processor.preProcess(fields, payload, req.files),
+      updated_at: new Date()
+    };
+
+    if (!payload._id) {
+      if (processor.checkExistence) {
+        model.findOne(processor.checkExistence(req)).then(record => {
+          if (record) {
+            const name = model.collection.collectionName.slice(0, -1);
+            res.status(400).json({errorReason: name + " already exists"});
+          } else {
+            doInsert(model, processor, update, res);
+          }
+        })
+      } else {
+        doInsert(model, processor, update, res);
+      }
+    }
+    // to update
+    else {
+      update._id = payload._id;
+      doUpdate(model, processor, update, res);
+    }
+  })
 };
 
 const defaultPaginationOptions = {
@@ -144,17 +155,25 @@ const getHandler = (model, processor) => (req, res) => {
 
 // only support delete one by id
 const deleteHandler = (model) => (req, res) => {
-  if (req.query._id) {
-    model
-      .findByIdAndRemove(mongoose.mongo.ObjectId(req.query._id),)
-      .then(data => {
-        res.json({ok: true, data: data});
-      }).catch((err) => {
-        res.status(400).json({"errReason": err});
-      });;
-  } else {
-    res.status(400).json({"errReason": "no id provided"});
-  }
+  const userId = req.user.id;
+  const propertyType = model.collection.collectionName;
+  const propertyId = req.query._id;
+  ACL.checkAccess(userId, propertyId, propertyType, "read_write", (isAccessible) => {
+    if (!isAccessible) {
+      return res.status(401).json({err: 'No permission to update.'});
+    }
+    if (req.query._id) {
+      model
+        .findByIdAndRemove(mongoose.mongo.ObjectId(req.query._id),)
+        .then(data => {
+          res.json({ok: true, data: data});
+        }).catch((err) => {
+          res.status(400).json({"errReason": err});
+        });
+    } else {
+      res.status(400).json({"errReason": "no id provided"});
+    }
+  });
 };
 
 module.exports = {
